chore: Update dependencies and install new packages
Update the chezmoiscripts to include new dependencies and install additional packages. This includes adding python, python-pip, python-pipx, and numix-circle-icon-theme to the pacman package list. It also installs xarchiver and its dependencies, and sassc for the wpgtk package. Finally, it updates the shell-gpt installation command to use pipx instead of pip3.