feat: Enable and start Docker service
This commit modifies the script `run_onchange_before_install-docker.sh.tmpl` to enable and start the Docker service using the `systemctl` command. By enabling and starting the Docker service, it ensures that Docker is automatically started on system boot and allows users to use Docker without manually starting the service. This improves the user experience and makes it more convenient to work with Docker.