feat: Add gvfs package for Thunar file manager
This commit adds the `gvfs` package to the pacman package list in the `run_onchange_before_install-thunar.sh.tmpl` script. The `gvfs` package provides support for virtual file systems in Thunar, allowing users to access remote files and devices seamlessly. This enhancement improves the functionality and usability of Thunar as a file manager.